WeaponBase
The WeaponBase class is an implementation of a generic base weapon for the game.
It's intended to be used as a parent class from which specific weapons can derive from. It provides basic characteristics common to all weapons such as rate of fire, reload speed, and the projectile it fires.

Key Features:
ProjectileClass: A UClass pointer used to specify the type of the projectile that the weapon fires.
RateOfFire: A float variable that represents the rate at which the weapon can fire projectiles. It is measured in x per second projectiles spawned.
ReloadSpeed: A float variable that represents the time passed for reloading the weapon. It is measured in seconds.