Update to WeaponBase
In the Collab09FPS.Build.cs file, the list of module names in PublicDependencyModuleNames is updated with some spacing correction, and new module names (including "GameplayAbilities", "GameplayTags", "GameplayTasks", and "Collab09FPS") are added to PrivateDependencyModuleNames.

In WeaponBase.cpp, the BeginPlay method is modified and three new methods - Fire, Melee, and Reload are added. The Fire method handles object spawning for the game, if the ProjectileClass is not null.
The WeaponBase.h file has received some updates:
It includes new imports for SkeletalMeshComponent.h, ArrowComponent.h, and ProjectileBase.h.
A new delegate FOnMelee is declared.
The ProjectileClass property is moved under the group "Weapon | Firing | Projectile".
A new UArrowComponent property, ProjectileSpawnLocation, is created under the group "Weapon | Firing | Projectile".
A new USkeletalMeshComponent property, WeaponMesh, is created under the group "Weapon".
The BaseRateOfFire and RateOfFire properties are moved under the group "Weapon | Firing |".
The BaseReloadSpeed and ReloadSpeed properties are moved under the group "Weapon | Firing |".
New properties for ammo count - MaxAmmoAmount and CurrentAmmoAmount - are introduced under the group "Weapon | Firing |".
The Fire, Melee, and Reload functions are declared.