Added checkpoints, re-added teleporters and kill zone, a minor adjustment in section one and a few concerns to raise!
The biggest addition I made is this one in section two. The first part is more of an indoors section (as they are meant to be hallways) with an extra obstacle to slide under.

The other key thing I added was the fans, as per my request. These work very well, the only thing I would add is the ability to adjust the strength of the wind they produce, so they don't push the player too much.

But there are a few concerns I'd like to raise. Firstly, while I have added a kill zone and a few checkpoints, for some reason I don't spawn at the checkpoint and instead end up at the tutorial level. Please let me know if this is something I am doing wrong or if the checkpoints simply need fixing.

Also, I am VERY certain that external actors has been checked this time (screenshot below as proof). If it doesn't work this time, I don't know what to do.