Starting to mesh block-out
So far, I've added the floor to section one of my level, will add the walls tomorrow and at the very least start the floor of section two.

While adding the meshes, I changed the material of the block-out stuff to a glowing blue to add contrast and thus make it easier to see the meshes as I added them. I also merged each group of floor instead of having hundreds of tiles.

Also, I added this part in section two as a way of introducing the fans mechanic in a way that doesn't put the player in immediate danger.