Starting to add extra details!
The main changes here are the wall running walls (to use the correct mesh) and a red tint across the whole level (as I saw the tutorial level do so).

The other thing to note is that the pink portals cannot be rotated. While this issue has already been raised, a second one is with the other portal type; it's invisible.