More decoration!
Mostly additions to Section Three. The main changes are the books (slightly edited the position of a few, and all of them are now textured) and I used some of the recently added props in Section Three's Arena to make it more of a dining section.

Still unsure of Section Two's theme, while Section One is meant to be more of a cathedral with plants and stuff.