Continued Decoration...
This time, I've added some more to section two (lights and storage props) and some plants in section one.

After the new imports, I also edited the pillars and the cupboards with books on them.

By this point, my priority is going to be on the sort of enemies that will be in my level and the sounds that might be needed as the decoration is pretty complete.

For some reason, Unreal kept crashing while I was working on this, do let me know if anyone else is experiencing the same.