Update to what my level will need
Changes to Level
I edited the strength of all the fans to be strong enough that they actually push the player, though some are stronger than others (the very first one is intentionally weaker because it serves as a way of introducing the mechanic.) Nothing else of note.
Sounds needed
•	Fans (A blowing sound)
•	Torches and candles (I’d imagine both will need similar burning sounds)
•	Electric barriers (If implemented, these will need an electric humming sound, the computer static sound which already exists may be perfect)
•	Lava (Probably needs a boiling sound so it’s not silent)
Enemies
Based on what the enemy folder contains, this is how they might be placed:
Section One (Before Arena): 9 enemies, all normal
Section One (At Arena): 6 enemies, introduce rangers (4 normal, 2 rangers)
Section Two (Before Arena): 10 enemies (7 normal, 3 rangers in the outside section)
Section Two (At Arena): 6 enemies, introduce DeepSeek
Section Three (Before Arena): 9 enemies (3 rangers and 2 DeepSeek in the elevated section, 4 normal before the first flight of stairs)
Section Three (At Arena): 10 enemies, introduce Heavy Rangers (4 Heavy Rangers in the elevated section to the sides, 2 normal rangers in the middle of the elevated section, 2 normal on each side of the lower section and two DeepSeek in the middle of the lower section)
It would be very helpful for me to have access to which enemies are being used exactly and their behaviours, in case any of the above is wrong. Either way, I have used red cones to mark where the enemies should be.
I will make a post in the design channel so that the rest of the design team is aware.